Move traits to their own directory.
